chore: standardize testing pipeline on pytest
- Added pytest-mock to requirements-test.txt.
- Refactored tests/conftest.py to move mock utilities to tests/mock_utils.py.
- Implemented autouse fixture for global mockfirestore patching in conftest.py.
- Updated tests/test_tournament.py and tests/test_tournament_blast.py to use the new mock_utils and removed explicit patch calls.
- Updated GitHub Actions workflows (main.yml and reusable-ci.yml) to use `uv run pytest` and include requirements-test.txt in installation.
