+++ .github/workflows/main.yml
@@ -39,7 +39,7 @@ jobs:
           python-version: 3.11
 
       - name: Install dependencies
-        run: uv pip install -r requirements.txt mock-firestore coverage --system
+        run: uv pip install -r requirements.txt -r requirements-test.txt coverage --system
 
       - name: Enable uuid-ossp extension
         run: |
@@ -54,8 +54,8 @@ jobs:
           POSTGRES_DB: postgres
           UV_SYSTEM_PYTHON: 1
         run: |
-          uv run --no-project python -m unittest discover tests
-          uv run --no-project python -m coverage run -m unittest discover tests
+          uv run --no-project python -m pytest
+          uv run --no-project python -m coverage run -m pytest
           uv run --no-project python -m coverage xml
 
       - name: Upload coverage to Codecov
