+++ .github/workflows/reusable-ci.yml
@@ -35,7 +35,7 @@ jobs:
           python-version: 3.11
 
       - name: Install dependencies
-        run: uv pip install -r requirements.txt mock-firestore coverage --system
+        run: uv pip install -r requirements.txt -r requirements-test.txt coverage --system
 
       - name: Enable uuid-ossp extension
         run: |
@@ -50,8 +50,8 @@ jobs:
           POSTGRES_DB: postgres
           UV_SYSTEM_PYTHON: 1
         run: |
-          uv run --no-project python -m unittest discover tests
-          uv run --no-project python -m coverage run -m unittest discover tests
+          uv run --no-project python -m pytest
+          uv run --no-project python -m coverage run -m pytest
           uv run --no-project python -m coverage xml
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v5
