+++ tests/test_tournament_blast.py
@@ -10,14 +10,11 @@ from mockfirestore import MockFirestore
 
 from pickaladder import create_app
 from pickaladder.user import UserService
-from tests.conftest import (
+from tests.mock_utils import (
     MockArrayRemove,
     MockArrayUnion,
     MockBatch,
-    patch_mockfirestore,
 )
-
-patch_mockfirestore()
 
 # Mock user payloads
 MOCK_USER_ID = "owner_id"
